Everything we discussed.
This is all the changes we had discussed for how this should ultimately look.

I'm very happy with this.
